Kyero.com Web Scraper by Python and Selenium_SourceAndOutput
This web scrapper gather all possible data from www.kyero.com listings. Due to a big piece of data are delivered by JS (ReacJS backend) and it is not possible to gather all this by just BeutifulSoup, the Python code is based on Selenium library. This is why this scrapper perform not in the best speed. The biggest delay is caused due to attempts to gather geographical (lat, lon) data from map script that is delivered by Google API.